update cart, products, productdetail, cartegory, user auth

+++ src/actions/auth.js
@@ -1,4 +1,4 @@
-import { publicRequest } from "../helpers/axios";
+import { publicRequest, userRequest } from "../helpers/axios";
 import {
     loginStart,
     loginSuccess,
@@ -7,36 +7,40 @@ import {
     logoutSuccess,
     signupStart,
     signupSuccess,
-    signupFailure
+    signupFailure,
+    refreshTokenStart,
+    refreshTokenSuccess,
+    refreshTokenFailure,
+    logoutFailure,
 } from "../redux/authRedux";
 
 export const signup = (user) => {
     return async (dispatch) => {
         dispatch(signupStart());
-        try{
+        try {
             const res = await publicRequest.post(`/api/signup`, user);
             console.log(res);
-            if(res.status === 201){
+            if (res.status === 201) {
                 dispatch(signupSuccess());
-            }else{
+            } else {
                 const { error } = res.data;
                 dispatch(signupFailure(error));
             }
-        }catch(err){
+        } catch (err) {
             const { data } = err.response;
-            dispatch(signupFailure({error: data.message}));
+            dispatch(signupFailure({ error: data.message }));
         }
-    }
-}
+    };
+};
 
 export const login = (user) => {
     return async (dispatch) => {
         dispatch(loginStart());
         try {
             const res = await publicRequest.post(`/api/signin`, user);
-            if(res.status === 200){
+            if (res.status === 200) {
                 dispatch(loginSuccess(res.data));
-            }else{
+            } else {
                 dispatch(loginFailure(res.data.error));
             }
         } catch (err) {
@@ -48,24 +52,37 @@ export const login = (user) => {
 export const logout = () => {
     return async (dispatch) => {
         dispatch(logoutStart());
-        localStorage.clear();
-        dispatch(logoutSuccess());
+        try {
+            const res = await userRequest.post(`/api/signout`);
+            if (res.status === 200) {
+                dispatch(logoutSuccess());
+                localStorage.clear();
+            } else {
+                dispatch(logoutFailure());
+            }
+        } catch (err) {
+            dispatch(logoutFailure());
+        }
     };
 };
 
 export const isUserLoggedin = () => {
     return async (dispatch) => {
-        const auth = JSON.parse(localStorage.getItem("persist:root"))?.auth;
-        const token = auth && JSON.parse(auth).token;
+        try {
+            const res = await userRequest.post("/api/isUserLoggedIn");
 
-        if (token) {
-            const user = auth && JSON.parse(auth).currentUser;
-            dispatch(
-                loginSuccess({
-                    token: token,
-                    user: user,
-                })
-            );
+            if (res.data.status === 401) {
+                dispatch(refreshTokenStart());  
+                const res = await userRequest.post("/api/refresh-token");
+                console.log(res)
+                if(res.data.status >= 300){
+                    dispatch(refreshTokenFailure());
+                }else{
+                    dispatch(refreshTokenSuccess(res.data));
+                }
+            }
+        } catch (err) {
+            dispatch(refreshTokenFailure());
         }
     };
 };
